perf: 优化代码，删除没有用到的代码
