+++ laokou-service/laokou-auth/laokou-auth-adapter/src/main/java/org/laokou/auth/consumer/handler/NoticeEventHandler.java
@@ -15,14 +15,17 @@
  *
  */
 
-package org.laokou.auth.consumer;
+package org.laokou.auth.consumer.handler;
 
 import io.micrometer.common.lang.NonNullApi;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.rocketmq.spring.annotation.RocketMQMessageListener;
+import org.laokou.auth.ability.DomainService;
+import org.laokou.auth.dto.domainevent.NoticeMessageEvent;
+import org.laokou.common.core.utils.JacksonUtil;
+import org.laokou.common.domain.handler.AbstractDomainEventHandler;
 import org.laokou.common.domain.support.DomainEventPublisher;
-import org.laokou.common.redis.utils.RedisUtil;
-import org.laokou.common.sms.service.SmsService;
+import org.laokou.common.i18n.dto.DefaultDomainEvent;
 import org.springframework.stereotype.Component;
 
 import static org.apache.rocketmq.spring.annotation.ConsumeMode.CONCURRENTLY;
@@ -35,24 +38,25 @@ import static org.laokou.auth.common.constant.MqConstant.*;
 @Slf4j
 @Component
 @NonNullApi
-@RocketMQMessageListener(consumerGroup = LAOKOU_MOBILE_CAPTCHA_CONSUMER_GROUP, topic = LAOKOU_CAPTCHA_TOPIC,
-		selectorExpression = MOBILE_TAG, messageModel = CLUSTERING, consumeMode = CONCURRENTLY)
-public class SendMobileCaptchaEventConsumer extends AbstractSendCaptchaEventConsumer {
+@RocketMQMessageListener(consumerGroup = LAOKOU_NOTICE_LOG_CONSUMER_GROUP, topic = LAOKOU_LOG_TOPIC,
+		selectorExpression = NOTICE_TAG, messageModel = CLUSTERING, consumeMode = CONCURRENTLY)
+public class NoticeEventHandler extends AbstractDomainEventHandler {
 
-	private final SmsService smsService;
+	private final DomainService domainService;
 
-	private final RedisUtil redisUtil;
-
-	public SendMobileCaptchaEventConsumer(DomainEventPublisher domainEventPublisher, SmsService smsService,
-			RedisUtil redisUtil) {
+	public NoticeEventHandler(DomainEventPublisher domainEventPublisher, DomainService domainService) {
 		super(domainEventPublisher);
-		this.smsService = smsService;
-		this.redisUtil = redisUtil;
+		this.domainService = domainService;
 	}
 
-	// @Override
-	// protected NoticeLog getNoticeLog(SendCaptchaEvent event) {
-	// return smsService.send(event.getUuid(), 5);
-	// }
+	@Override
+	protected void handleDomainEvent(DefaultDomainEvent domainEvent) {
+		// authDomainService.recordNoticeLog(domainEvent);
+	}
+
+	@Override
+	protected DefaultDomainEvent convert(String msg) {
+		return JacksonUtil.toBean(msg, NoticeMessageEvent.class);
+	}
 
 }
