+++ laokou-service/laokou-auth/laokou-auth-app/src/main/java/org/laokou/auth/service/extensionpoint/extension/MailAuthParamValidator.java
@@ -22,8 +22,8 @@ import org.laokou.auth.model.AuthA;
 import org.laokou.common.core.utils.RegexUtil;
 import org.laokou.common.extension.Extension;
 
-import static org.laokou.common.domain.utils.ParamValidatorUtil.validateNotEmpty;
-import static org.laokou.common.domain.utils.ParamValidatorUtil.validateRegex;
+import static org.laokou.auth.common.utils.ParamValidatorUtil.validateNotEmpty;
+import static org.laokou.auth.common.utils.ParamValidatorUtil.validateRegex;
 import static org.laokou.auth.factory.DomainFactory.MAIL;
 import static org.laokou.auth.model.AuthA.USE_CASE_AUTH;
 import static org.laokou.common.i18n.common.constant.Constant.SCENARIO;
