+++ laokou-service/laokou-auth/laokou-auth-app/src/main/java/org/laokou/auth/service/extensionpoint/extension/MailCaptchaParamValidator.java
@@ -22,7 +22,7 @@ import org.laokou.common.core.utils.RegexUtil;
 import org.laokou.common.extension.Extension;
 
 import static org.laokou.auth.common.constant.MqConstant.MAIL_TAG;
-import static org.laokou.common.domain.utils.ParamValidatorUtil.validateRegex;
+import static org.laokou.auth.common.utils.ParamValidatorUtil.validateRegex;
 import static org.laokou.auth.dto.CaptchaSendCmd.USE_CASE_CAPTCHA;
 import static org.laokou.common.i18n.common.constant.Constant.SCENARIO;
 import static org.laokou.common.i18n.common.exception.ParamException.OAuth2.MAIL_ERROR;
