+++ laokou-service/laokou-auth/laokou-auth-app/src/main/java/org/laokou/auth/service/extensionpoint/extension/PasswordAuthParamValidator.java
@@ -21,7 +21,7 @@ import org.laokou.auth.service.extensionpoint.AuthParamValidatorExtPt;
 import org.laokou.auth.model.AuthA;
 import org.laokou.common.extension.Extension;
 
-import static org.laokou.common.domain.utils.ParamValidatorUtil.validateNotEmpty;
+import static org.laokou.auth.common.utils.ParamValidatorUtil.validateNotEmpty;
 import static org.laokou.auth.factory.DomainFactory.PASSWORD;
 import static org.laokou.auth.model.AuthA.USE_CASE_AUTH;
 import static org.laokou.common.i18n.common.constant.Constant.SCENARIO;
